requirements.txt
textual>=0.47.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Note: You'll also need these dependencies from your original script:
# - py-ankiconnect
# - google-generativeai (for gemini)
# - google-cloud-texttospeech (for tts)
# 
# Install with:
# pip install py-ankiconnect google-generativeai google-cloud-texttospeech
